Adicionado recursos basicos do backend para gerenciamento de níveis

+++ dev-manager-api/src/nivel/nivel.controller.ts
@@ -1,15 +1,25 @@
-import { Controller, Get, Post, Body, Patch, Param, Delete } from '@nestjs/common';
+import { Controller, Get, Post, Body, Patch, Param, Delete, Res, HttpStatus, Put } from '@nestjs/common';
 import { NivelService } from './nivel.service';
 import { CreateNivelDto } from './dto/create-nivel.dto';
 import { UpdateNivelDto } from './dto/update-nivel.dto';
+import { validate } from 'class-validator';
+import { Response } from 'express';
 
 @Controller('nivel')
 export class NivelController {
   constructor(private readonly nivelService: NivelService) {}
 
   @Post()
-  create(@Body() createNivelDto: CreateNivelDto) {
-    return this.nivelService.create(createNivelDto);
+  async create(@Body() createNivelDto: CreateNivelDto, @Res() res: Response): Promise<any>  {
+    try {
+      const error = await validate(this.nivelService.nivelFromDTO(createNivelDto));
+      if (error.length > 0) {
+        return res.status(HttpStatus.BAD_REQUEST).send(error);
+      }
+      return res.status(HttpStatus.CREATED).json(await this.nivelService.create(createNivelDto));
+    } catch (e) {
+      return res.status(HttpStatus.BAD_REQUEST).send(e);
+    }
   }
 
   @Get()
@@ -22,13 +32,39 @@ export class NivelController {
     return this.nivelService.findOne(+id);
   }
 
+  @Put(':id')
+  async update(@Param('id') id: string, @Body() updateNivelDto: UpdateNivelDto, @Res() res: Response): Promise<any> {
+    const nivel = await this.nivelService.findOne(+id);
+    if (nivel){
+      try{
+        const nivel = await this.nivelService.update(+id, updateNivelDto);
+        return res.status(HttpStatus.OK).json(nivel);
+      } catch (error) {
+        return res.status(HttpStatus.BAD_REQUEST).send(error);
+      }
+    } else {
+      return res.status(HttpStatus.NOT_FOUND).send("Nivel com id " + +id + " não encontrado!");  
+    }
+  }
+
   @Patch(':id')
-  update(@Param('id') id: string, @Body() updateNivelDto: UpdateNivelDto) {
-    return this.nivelService.update(+id, updateNivelDto);
+  async patch(@Param('id') id: string, @Body() updateNivelDto: UpdateNivelDto, @Res() res: Response) {
+    return await this.update(id, updateNivelDto, res);
   }
 
   @Delete(':id')
-  remove(@Param('id') id: string) {
-    return this.nivelService.remove(+id);
+  async remove(@Param('id') id: string, @Res() res: Response) {
+    const nivel = await this.nivelService.findOne(+id);
+    if (nivel){    
+      try {
+        await this.nivelService.remove(+id);
+        return res.status(HttpStatus.NO_CONTENT).send();
+      } catch (error) {
+        console.log(error);
+        return res.status(HttpStatus.BAD_REQUEST).json(error);
+      }
+    } else {
+      return res.status(HttpStatus.NOT_FOUND).send("Nivel com id " + +id + " não encontrado!");  
+    }
   }
 }
